moved most of API module to commons
Builds, but following error message occurs when simulating android app:
E/FirebaseCrashlytics: Settings request failed.
    java.io.FileNotFoundException: https://firebase-settings.crashlytics.com/spi/v2/platforms/android/gmp/1:460393230596:android:a4512316c4b07001132051/settings?instance=d2ce15d64f057f5a768d8a0187b20776f9e4e11b&build_version=581&display_version=0.7.0&source=1
