--- app/src/main/java/hpsaturn/pollutionreporter/api/AqicnIaqi.java
@@ -1,19 +1,0 @@
-package hpsaturn.pollutionreporter.api;
-
-/**
- * Created by Antonio Vanegas @hpsaturn on 12/28/19.
- */
-public class AqicnIaqi {
-
-    public AqicnCo co;
-    public AqicnH h;
-    public AqicnNo2 no2;
-    public AqicnO3 o3;
-    public AqicnP p;
-    public AqicnPm10 pm10;
-    public AqicnPm25 pm25;
-    public AqicnSo2 so2;
-    public AqicnT t;
-    public AqicnW w;
-
-}
