+++ commons/src/main/java/com/jetbrains/handson/commons/api/AqicnAttributions.java
@@ -1,4 +1,4 @@
-package hpsaturn.pollutionreporter.api;
+package com.jetbrains.handson.commons.api;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 12/28/19.
