+++ commons/src/main/java/com/jetbrains/handson/commons/api/AqicnCity.java
@@ -1,11 +1,11 @@
-package hpsaturn.pollutionreporter.api;
+package com.jetbrains.handson.commons.api;
 
 import java.util.List;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 12/28/19.
  */
-class AqicnCity {
+public class AqicnCity {
 
     public String name;
