+++ commons/src/main/java/com/jetbrains/handson/commons/api/AqicnDataResponse.java
@@ -1,4 +1,6 @@
-package hpsaturn.pollutionreporter.api;
+package com.jetbrains.handson.commons.api;
+
+import com.jetbrains.handson.commons.api.AqicnData;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 12/28/19.
