+++ commons/src/main/java/com/jetbrains/handson/commons/api/AqicnH.java
@@ -1,10 +1,9 @@
-package hpsaturn.pollutionreporter.api;
+package com.jetbrains.handson.commons.api;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 12/28/19.
  */
-public class AqicnSo2 {
+public class AqicnH {
 
     public double v;
-
 }
