+++ commons/src/main/java/com/jetbrains/handson/commons/api/AqicnIaqi.java
@@ -1,0 +1,30 @@
+package com.jetbrains.handson.commons.api;
+
+import com.jetbrains.handson.commons.api.AqicnCo;
+import com.jetbrains.handson.commons.api.AqicnH;
+import com.jetbrains.handson.commons.api.AqicnNo2;
+import com.jetbrains.handson.commons.api.AqicnO3;
+import com.jetbrains.handson.commons.api.AqicnP;
+import com.jetbrains.handson.commons.api.AqicnPm10;
+import com.jetbrains.handson.commons.api.AqicnPm25;
+import com.jetbrains.handson.commons.api.AqicnSo2;
+import com.jetbrains.handson.commons.api.AqicnT;
+import com.jetbrains.handson.commons.api.AqicnW;
+
+/**
+ * Created by Antonio Vanegas @hpsaturn on 12/28/19.
+ */
+public class AqicnIaqi {
+
+    public AqicnCo co;
+    public AqicnH h;
+    public AqicnNo2 no2;
+    public AqicnO3 o3;
+    public AqicnP p;
+    public AqicnPm10 pm10;
+    public AqicnPm25 pm25;
+    public AqicnSo2 so2;
+    public AqicnT t;
+    public AqicnW w;
+
+}
