+++ commons/src/main/java/com/jetbrains/handson/commons/api/AqicnO3.java
@@ -1,9 +1,9 @@
-package hpsaturn.pollutionreporter.api;
+package com.jetbrains.handson.commons.api;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 12/28/19.
  */
-public class AqicnP {
+public class AqicnO3 {
 
     public double v;
 }
