+++ commons/src/main/java/com/jetbrains/handson/commons/api/AqicnSo2.java
@@ -1,9 +1,10 @@
-package hpsaturn.pollutionreporter.api;
+package com.jetbrains.handson.commons.api;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 12/28/19.
  */
-public class AqicnNo2 {
+public class AqicnSo2 {
 
     public double v;
+
 }
